requirements.txt
# Product Enrichment Pipeline Dependencies

# Core libraries
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# AI/API clients
anthropic>=0.18.0

# Utilities
urllib3>=2.0.0
